+++ packages/terra-graphs/tests/wdio/Graphs/LineGraph/line-graph-spec.js
@@ -166,4 +166,118 @@ Terra.describeViewports('LineGraph ', ['tiny', 'medium', 'large'], () => {
       Terra.validates.screenshot('simple_spline', { selector: '.carbon-graph-container' });
     });
   });
+
+  describe('Reflow', () => {
+    describe('simple panning', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/simple-panning');
+      });
+      it('validates panning left', () => {
+        panningBtn = $('.button-pan-left');
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_3', { selector: '.carbon-graph-container' });
+      });
+      it('validates panning right', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_3', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('dynamic data', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/dynamic-data');
+      });
+      it('validates dynaimcally updating of a dataset', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_data_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_data_2', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('multiple datasets', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/multiple-datasets');
+      });
+      it('validates dynaimcally updating of multiple datasets', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('multiple_dataset_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('multiple_dataset_2', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('dynamic eventline', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/dynamic-eventline');
+      });
+      it('validates eventline dynaimcally updating', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_eventline_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_eventline_2', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('legend update', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/dynamic-legend');
+      });
+      it('validates eventline dynaimcally updating', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('legend_update_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('legend_update_2', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('y2 panning', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/terra-graphs-docs/graphs/line-graph/reflow/y-2-panning');
+      });
+      it('validates panning left with y2 axis', () => {
+        panningBtn = $('.button-pan-left');
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_left_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_left_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_left_3', { selector: '.carbon-graph-container' });
+      });
+      it('validates panning right with y2 axis', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_right_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_right_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('y2_panning_right_3', { selector: '.carbon-graph-container' });
+      });
+    });
+  });
 });