+++ packages/terra-graphs/CHANGELOG.md
@@ -11,6 +11,7 @@
   * Created component and wdio test cases for the Timeline graph.
   * Created component and wdio test cases for the Pie chart.
   * Created component and wdio test cases for the Combination graph.
+  * Added reflow to the Line component.
   * Added CSS and new example tests for timeline graph.
 
 * Initial stable release